requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-generativeai
torch
torchvision
ultralytics
numpy
Pillow
opencv-python-headless